feat: add lib.ts for judge/mentor registration function
