+++ app/registration/page.tsx
@@ -1,13 +1,11 @@
-import { getFirestore } from "firebase-admin/firestore";
 import { redirect } from "next/navigation";
 
-import { DASHBOARD_PATH, JUDGE, LOGIN_PATH, MENTOR, USERS_COLLECTION } from "@/constants";
+import { LOGIN_PATH } from "@/constants";
 import { getConfigDocSnapshot, verifySession } from "@/lib";
-import type { JudgeUser, WildHacksConfig } from "@/types";
+import type { WildHacksConfig } from "@/types";
 
 import RegistrationForm from "./_components/registration-form";
-
-const getCurrentTimestamp = () => Date.now();
+import { registerJudgeMentorWithEmail } from "./lib";
 
 const RegistrationPage = async () => {
   const userInfo = await verifySession();
@@ -15,32 +13,7 @@ const RegistrationPage = async () => {
 
   const { id, email } = userInfo;
 
-  const db = getFirestore();
-  const userDocSnapshotById = await db.collection(USERS_COLLECTION).doc(id).get();
-  if (userDocSnapshotById.exists) redirect(DASHBOARD_PATH);
-
-  const userDocSnapshotByEmail = await db.collection(USERS_COLLECTION).where("email", "==", email).limit(1).get();
-  if (!userDocSnapshotByEmail.empty) {
-    const newJudgeDocRef = db.collection(USERS_COLLECTION).doc(id);
-
-    const data = userDocSnapshotByEmail.docs[0].data();
-
-    if (data?.role === JUDGE || data?.role === MENTOR) {
-      const timestamp = getCurrentTimestamp();
-
-      await db.runTransaction(async (transaction) => {
-        transaction.set(newJudgeDocRef, {
-          ...data,
-          created_at: timestamp,
-          updated_at: timestamp,
-        } as JudgeUser);
-
-        transaction.delete(userDocSnapshotByEmail.docs[0].ref);
-      });
-
-      redirect(DASHBOARD_PATH);
-    }
-  }
+  await registerJudgeMentorWithEmail(id, email);
 
   const configDocSnapshot = await getConfigDocSnapshot();
   const wildhacksConfig = configDocSnapshot.data() as WildHacksConfig;
